+++ org-slipbox.el
@@ -31,6 +31,7 @@
 
 ;;; Code:
 
+(require 'org-slipbox-agenda)
 (require 'org-slipbox-node)
 (require 'org-slipbox-rpc)
 (require 'org-slipbox-dailies)
@@ -51,6 +52,8 @@
 ;;;###autoload(autoload 'org-slipbox-node-find "org-slipbox-node" nil t)
 ;;;###autoload(autoload 'org-slipbox-node-insert "org-slipbox-node" nil t)
 ;;;###autoload(autoload 'org-slipbox-node-backlinks "org-slipbox-node" nil t)
+;;;###autoload(autoload 'org-slipbox-agenda-today "org-slipbox-agenda" nil t)
+;;;###autoload(autoload 'org-slipbox-agenda-date "org-slipbox-agenda" nil t)
 ;;;###autoload(autoload 'org-slipbox-dailies-capture-today "org-slipbox-dailies" nil t)
 ;;;###autoload(autoload 'org-slipbox-dailies-goto-today "org-slipbox-dailies" nil t)
 ;;;###autoload(autoload 'org-slipbox-dailies-capture-tomorrow "org-slipbox-dailies" nil t)